Add Pocos to shared library (moved from API.models)
