+++ DailyPmsAPI/DailyPmsAPI.csproj
@@ -8,7 +8,7 @@
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <WarningLevel>4</WarningLevel>
     <DocumentationFile>bin\Debug\net5.0\DailyPmsAPI.xml</DocumentationFile>
-    <NoWarn>1591,1702</NoWarn>
+    <NoWarn>1591,1702,1701</NoWarn>
   </PropertyGroup>
   <ItemGroup>
     <Folder Include="Controllers\" />
@@ -18,5 +18,10 @@
   <ItemGroup>
     <PackageReference Include="MongoDB.Driver" Version="2.12.1" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.1.1" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\DailyPmsShared\DailyPmsShared.csproj">
+      <GlobalPropertiesToRemove></GlobalPropertiesToRemove>
+    </ProjectReference>
   </ItemGroup>
 </Project>
