+++ DailyPmsAPI/Data/IDbContext.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using MongoDB.Driver;
-using DailyPmsAPI.Models;
+using DailyPmsShared;
 
 namespace DailyPmsAPI.Data
 {
