+++ DailyPmsAPI/Data/IPmsFileRepository.cs
@@ -1,5 +1,5 @@
 ﻿using System.Threading.Tasks;
-using DailyPmsAPI.Models;
+using DailyPmsShared;
 
 namespace DailyPmsAPI.Data
 {
