+++ DailyPmsAPI/Data/MongoDbContext.cs
@@ -1,5 +1,5 @@
 ﻿using MongoDB.Driver;
-using DailyPmsAPI.Models;
+using DailyPmsShared;
 
 namespace DailyPmsAPI.Data
 {
