+++ DailyPmsAPI/Data/MongoPmsFileRepository.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
-using DailyPmsAPI.Models;
+using DailyPmsShared;
 using MongoDB.Driver;
 
 namespace DailyPmsAPI.Data
